ci: Renamed package name to exclude conflicts
